test(pass): test ub and ddr, test ddr

+++ framework/src/passes/tile_graph_pass/graph_constraint/pre_graph/pre_graph.cpp
@@ -71,8 +71,10 @@ Status PreGraphProcess::RunOnFunction(Function &function) {
             setCopyAttr.ProcessMoveInOperation(op);
         }
     }
-    RemoveRedundantAssemble removeRedundantAssemble;
-    removeRedundantAssemble.DeleteRedundantAssemble(function);
+    if (function.paramConfigs_.perGraphAssRemove) {
+        RemoveRedundantAssemble removeRedundantAssemble;
+        removeRedundantAssemble.DeleteRedundantAssemble(function);
+    }
     CubeProcess cubeProcess;
     if (cubeProcess.UpdateCubeOp(function) != SUCCESS) {
         APASS_LOG_ERROR_F(Elements::Function, "Update Cube attr failed.");
